Cast volatile currentStatus when calling setValue
The volatile qualifier on currentStatus prevents implicit conversion
in the setValue template. Cast to plain uint8_t at the call site.

+++ src/main.cpp
@@ -97,7 +97,7 @@ void setup() {
         NIMBLE_PROPERTY::READ | NIMBLE_PROPERTY::WRITE
     );
     pStatusChar->setCallbacks(&charCallbacks);
-    pStatusChar->setValue(currentStatus);
+    pStatusChar->setValue((uint8_t)currentStatus);
     pService->start();
 
     NimBLEAdvertising* pAdvertising = NimBLEDevice::getAdvertising();
